refactor: Move consensus::*::tests into their own files
