+++ zebra-consensus/src/redjubjub.rs
@@ -1,5 +1,8 @@
 //! Async RedJubjub batch verifier service
 
+#[cfg(test)]
+mod tests;
+
 use std::{
     future::Future,
     mem,
@@ -81,75 +84,5 @@ impl Drop for Verifier {
         // We need to flush the current batch in case there are still any pending futures.
         let batch = mem::take(&mut self.batch);
         let _ = self.tx.send(batch.verify(thread_rng()));
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    use std::time::Duration;
-
-    use color_eyre::eyre::Result;
-    use futures::stream::{FuturesUnordered, StreamExt};
-    use tower::ServiceExt;
-    use tower_batch::Batch;
-
-    async fn sign_and_verify<V>(mut verifier: V, n: usize) -> Result<(), V::Error>
-    where
-        V: Service<Item, Response = ()>,
-    {
-        let rng = thread_rng();
-        let mut results = FuturesUnordered::new();
-        for i in 0..n {
-            let span = tracing::trace_span!("sig", i);
-            let msg = b"BatchVerifyTest";
-
-            match i % 2 {
-                0 => {
-                    let sk = SigningKey::<SpendAuth>::new(rng);
-                    let vk = VerificationKey::from(&sk);
-                    let sig = sk.sign(rng, &msg[..]);
-                    verifier.ready_and().await?;
-                    results.push(span.in_scope(|| verifier.call((vk.into(), sig, msg).into())))
-                }
-                1 => {
-                    let sk = SigningKey::<Binding>::new(rng);
-                    let vk = VerificationKey::from(&sk);
-                    let sig = sk.sign(rng, &msg[..]);
-                    verifier.ready_and().await?;
-                    results.push(span.in_scope(|| verifier.call((vk.into(), sig, msg).into())))
-                }
-                _ => panic!(),
-            }
-        }
-
-        while let Some(result) = results.next().await {
-            result?;
-        }
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    #[spandoc::spandoc]
-    async fn batch_flushes_on_max_items() -> Result<()> {
-        use tokio::time::timeout;
-
-        // Use a very long max_latency and a short timeout to check that
-        // flushing is happening based on hitting max_items.
-        let verifier = Batch::new(Verifier::new(), 10, Duration::from_secs(1000));
-        timeout(Duration::from_secs(5), sign_and_verify(verifier, 100)).await?
-    }
-
-    #[tokio::test]
-    #[spandoc::spandoc]
-    async fn batch_flushes_on_max_latency() -> Result<()> {
-        use tokio::time::timeout;
-
-        // Use a very high max_items and a short timeout to check that
-        // flushing is happening based on hitting max_latency.
-        let verifier = Batch::new(Verifier::new(), 100, Duration::from_millis(500));
-        timeout(Duration::from_secs(5), sign_and_verify(verifier, 10)).await?
     }
 }
